[inbox] show message states + allow retry
